3 - zones: extract logic to open URLs from zones
